refactor: move database client to tests/common
