+++ src/meta-srv/Cargo.toml
@@ -61,7 +61,7 @@ url = "2.3"
 
 [dev-dependencies]
 chrono.workspace = true
-client = { workspace = true, features = ["testing"] }
+client.workspace = true
 common-meta = { workspace = true, features = ["testing"] }
 common-procedure-test.workspace = true
 session.workspace = true
