+++ tests-integration/Cargo.toml
@@ -69,6 +69,7 @@ sqlx = { version = "0.6", features = [
 substrait.workspace = true
 table.workspace = true
 tempfile.workspace = true
+tests-common.workspace = true
 time = "0.3"
 tokio.workspace = true
 tokio-stream = { workspace = true, features = ["net"] }
