+++ tests/common/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "tests-common"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+api.workspace = true
+arrow-flight.workspace = true
+async-stream.workspace = true
+async-trait = "0.1"
+client.workspace = true
+common-error.workspace = true
+common-grpc.workspace = true
+common-query.workspace = true
+common-recordbatch.workspace = true
+common-telemetry.workspace = true
+futures.workspace = true
+futures-util.workspace = true
+prost.workspace = true
+snafu.workspace = true
+tonic.workspace = true
+
+[dev-dependencies]
+clap.workspace = true
+cmd.workspace = true
+common-catalog.workspace = true
+common-telemetry.workspace = true
+tempfile.workspace = true
+tokio.workspace = true
