+++ tests/common/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2023 Greptime Team
+// Copyright 2024 Greptime Team
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,22 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![feature(assert_matches)]
-
-pub mod cluster;
-mod grpc;
-mod influxdb;
-mod instance;
-mod opentsdb;
-mod otlp;
-mod prom_store;
-pub mod test_util;
-
-pub mod standalone;
-#[cfg(test)]
-mod tests;
-
-#[cfg(test)]
-// allowed because https://docs.rs/rstest_reuse/0.5.0/rstest_reuse/#use-rstest_reuse-at-the-top-of-your-crate
-#[allow(clippy::single_component_path_imports)]
-use rstest_reuse;
+pub mod database;
