+++ tests/runner/Cargo.toml
@@ -19,7 +19,6 @@ serde.workspace = true
 serde_json.workspace = true
 sqlness = { version = "0.5" }
 tempfile.workspace = true
-# TODO depend `Database` client
-tests-integration.workspace = true
+tests-common.workspace = true
 tinytemplate = "1.2"
 tokio.workspace = true
